third_party/blink/web_tests/fast/xpath/xpath-functional-test-expected.txt

[ok].//blockquote/*
[ok].//blockquote/child::*
[ok].//blockquote/parent::*
[ok].//blockquote/descendant::*
[ok].//blockquote/descendant-or-self::*
[ok].//blockquote/ancestor::*
[ok].//blockquote/ancestor-or-self::*
[ok].//blockquote/following-sibling::*
[ok].//blockquote/preceding-sibling::*
[ok].//blockquote/following::*
[ok].//blockquote/preceding::*
[ok].//blockquote/self::*
[ok].//blockquote/attribute::id/parent::*
[ok].//blockquote/@id/parent::*
[ok].//*[blockquote]
[ok].//*[child::blockquote]
[ok].//*[parent::blockquote]
[ok].//*[descendant::blockquote]
[ok].//*[descendant-or-self::blockquote]
[ok].//*[ancestor::blockquote]
[ok].//*[ancestor-or-self::blockquote]
[ok].//*[following-sibling::blockquote]
[ok].//*[preceding-sibling::blockquote]
[ok].//*[following::blockquote]
[ok].//*[preceding::blockquote]
[ok].//*[self::blockquote]
[ok].//*[@id]
[ok].//*[attribute::id]
[ok].//blockquote/text()
[ng].//blockquote/comment()
[ng].//blockquote/processing-instruction()
[ng].//blockquote/processing-instruction("pi")
[ng].//blockquote/node()
[ok].//blockquote/p
[ok].//blockquote/*
[ok].//*[child::* and preceding::font]
[ok].//*[not(child::*) and preceding::font]
[ok].//*[preceding::blockquote or following::blockquote]
[ok].//blockquote/ancestor::* | .//blockquote/descendant::*
[ok].//*[.="sub"]
[ok].//*[@title > 12 and @class < 15]
[ok].//*[@title != @class]
[ok].//*[((@class * @class + @title * @title) div (@class + @title)) > ((@class - @title) * (@class - @title))]
[ok].//*[@title mod 2 = 0]
[ok].//blockquote/child::*[last()]
[ok].//blockquote/descendant::*[position() < 4]
[ok]id(.//font/@face)
[ok].//*[name(.) = "sub"]
[ok].//*[name() = "sub"]
[ok].//blockquote/child::*[2]
[ok].//blockquote/descendant::*[4]
[ok].//blockquote/descendant-or-self::*[4]
[ok].//blockquote/ancestor::*[2]
[ok].//blockquote/ancestor-or-self::*[2]
[ok].//blockquote/following-sibling::*[1]
[ok].//blockquote/preceding-sibling::*[1]
[ok].//blockquote/following::*[4]
[ok].//blockquote/preceding::*[4]
[ok].//*[starts-with(.,"s")]
[ok].//*[string(@title - 1) = "0"]
[ok].//*[string() = "sub"]
[ok].//*[string(.) = "sub"]
[ok].//*[concat(.,..) = "subsubsup"]
[ok].//node()[concat(.,..,../..) = "bbbs"]
[ok].//*[starts-with(.,"s")]
[ok].//*[substring-before(.,"u") = "s"]
[ok].//*[substring-after(.,"on") = "t"]
[ok].//*[substring(.,2,1) = "u"]
[ok].//*[substring(.,2) = "up"]
[ok].//*[contains(.,"b")]
[ok].//*[string-length() = 3]
[ok].//*[string-length(.) = 3]
[ok].//*[.=translate(normalize-space(" s u b ")," ","")]
[ok].//*[normalize-space()="q"]
[ok].//*[boolean(@title - 1) = false()]
[ok].//*[not(@title - 1) = true()]
[ng].//*[lang("it")]
[ok].//*[number(@title) < number(@class)]
[ok].//*[sum(ancestor::*/@title) < sum(descendant::*/@title)]
[ok].//*[floor(@title div @class) = 1]
[ok].//*[ceiling(@title div @class) = 1]
[ok].//*[round(@title div @class) = 1]
[ok].//*[blockquote]
dt
dd
emstrong
bs
blockquoteText1:
blockquoteText2

delins

font
dfna
subsup
abbrq
